group-dps: refactor to accept decimal values && enhance the visual aspect of the calculator with css

+++ GroupDPS/javaScript/task/calculator/main.js
@@ -19,11 +19,11 @@ let resultArea = document.getElementById("result");
 // function definitions
 function setFirstNum(event) {
   // x => something
-  x = parseInt(event.target.value);
+  x = parseFloat(event.target.value);
   console.log(x);
 }
 function setSecondNum(event) {
-  y = parseInt(event.target.value);
+  y = parseFloat(event.target.value);
   console.log(y);
 }
